refs(api): Stop returning unique values seen from most tag endpoints.
Based on https://github.com/getsentry/sentry/pull/12115. This will give us a substantial performance
improvement when fetching tags. The `uniqueValues` value isn't used at all by our frontend, and so
it should be fine to just remove it completely.

There are some old endpoints where someone might have an integration with us that this breaks.
`ProjectTagsEndpoint` isn't a concern for performance, so I've left it as is for now.
`GroupTagsEndpoint` could potentially cause issues, but is likely worthwhile to take the risk for
the performance gain. `GroupDetailsEndpoint` is in a similar spot to `GroupTagsEndpoint`, probably
worth the risk.

If we want to be more cautious, we could parameterize the endpoints and have FE pass in a value so
that we stop including this value.

Fixes SEN-235